services/vanna/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
vanna==0.3.4
psycopg2-binary==2.9.9
python-dotenv==1.0.0
groq==0.4.1